removed sign up view and replaced with sign up modal
